+++ src/template/__init__.py
@@ -1,3 +1,0 @@
-""" Template """
-
-__version__ = "0.1.0"