+++ src/template/template.py
@@ -1,6 +1,0 @@
-class Template(object):
-    def __init__(self):
-        pass
-
-    def sum(self, value1, value2):
-        return value1 + value2