+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# 2024.02.23
+# Changelog
 
-- Features: Create this template with a sum function
+## v0.1.1 2024.03.14
+
+- Rename packages from `template` to `templatelib`.
+- Bugfix: Convert string values to float.
+
+## v0.1.0 2024.02.23
+
+- Features: Create this template with a sum function.
 - Bugfix: ...